requirements.txt
# Core dependencies
SpeechRecognition>=3.10.0
PyAudio>=0.2.13
python-dotenv>=1.0.0
openai>=1.12.0
google-cloud-speech>=2.21.3
requests>=2.32.2
psutil>=5.9.8  # Nueva dependencia para monitoreo de recursos

# Text-to-Speech
gTTS>=2.4.0
pygame>=2.5.2

# LLM Local
llama-cpp-python>=0.2.61  # Actualizado para soporte de nuevos modelos

# Modelos LLM
google-generativeai>=0.3.2  # Para Gemini

# Progreso de descargas
tqdm>=4.66.2  # Nueva dependencia para barras de progreso

# Gestión de modelos
numpy>=1.26.4  # Dependencia implícita para varios componentes